+++ apps/api/src/modules/squad-builder/dto/save-squad.dto.ts
@@ -1,0 +1,7 @@
+import { IsString } from 'class-validator';
+import { GenerateSquadDto } from './generate-squad.dto';
+
+export class SaveSquadDto extends GenerateSquadDto {
+  @IsString()
+  name!: string;
+}